pyproject.toml
[project]
name = "swarden"
version = "1.0.0"
description = "Add your description here"
readme = { file = ".github/README.md", content-type = "text/markdown" }
license = { file = "LICENSE" }
authors = [{ name = "LucasGonc", email = "lucasgoncsilva04@gmail.com" }]
maintainers = [{ name = "LucasGonc", email = "lucasgoncsilva04@gmail.com" }]


requires-python = ">=3.13"
dependencies = [
    "crispy-bootstrap4>=2024.10",
    "dj-database-url>=2.3.0",
    "django>=5.1.5",
    "django-crispy-forms>=2.3",
    "django-monthfield>=0.1.4",
    "django-ranged-response>=0.2.0",
    "django-simple-captcha>=0.6.1",
    "gunicorn>=23.0.0",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "whitenoise>=6.8.2",
]

[dependency-groups]
dev = [
    "coverage>=7.6.10",
    "django-extensions>=3.2.3",
    "locust>=2.32.6",
    "mosheh>=1.3.4",
    "ruff>=0.9.2",
    "tblib>=3.0.0",
    "tqdm>=4.67.1",
]


[project.urls]
Documentation = "https://lucasgoncsilva.github.io/swarden/"
Repository = "https://github.com/lucasGoncSilva/swarden/"
Issues = "https://github.com/LucasGoncSilva/swarden/issues"
Changelog = "https://github.com/LucasGoncSilva/swarden/blob/main/.github/CHANGELOG.md"


[tool.ruff]
exclude = [".env", "migrations", "SWARDEN/secret/month"]
fix = true

[tool.ruff.format]
quote-style = "single"
docstring-code-format = true

[tool.ruff.lint]
select = ["I", "E", "UP", "F"]
fixable = ["ALL"]
exclude = [".env", "migrations", "SWARDEN/secret/month"]

[tool.ruff.lint.isort]
case-sensitive = true
lines-after-imports = 2

[tool.ruff.lint.pycodestyle]
max-doc-length = 88
